Fix some ci and local ci issues
